% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_dubois2.R
\docType{data}
\name{colors2}
\alias{colors2}
\title{Short color palette (fill)}
\format{
An object of class \code{character} of length 3.
}
\usage{
colors2
}
\value{
Fill with short color palette
}
\description{
Short color palette (fill)
}
\keyword{datasets}
