% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_dubois1.R
\name{scale_color_dubois1}
\alias{scale_color_dubois1}
\title{Long color palette (color)}
\usage{
scale_color_dubois1()
}
\value{
Color with long color palette
}
\description{
Long color palette (color)
}
