% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_dubois2.R
\name{scale_color_dubois2}
\alias{scale_color_dubois2}
\title{Short color palette (color)}
\usage{
scale_color_dubois2()
}
\value{
Color with short color palette
}
\description{
Short color palette (color)
}
