% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_dubois2.R
\name{scale_fill_dubois2}
\alias{scale_fill_dubois2}
\title{Short color palette (fill)}
\usage{
scale_fill_dubois2()
}
\value{
Fill with short color palette
}
\description{
Short color palette (fill)
}
