% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dubois.R
\name{theme_dubois}
\alias{theme_dubois}
\title{A custom theme based on W.E.B. Du Bois' visualizations}
\usage{
theme_dubois()
}
\value{
A ggplot2 plot
}
\description{
A custom theme based on W.E.B. Du Bois' visualizations
}
\examples{
For documentattion and examples please visit https://github.com/vladmedenica/themedubois
}
